Use World-owned SSE trace buffer
Replace the SSE BDD test's dependency on `tracing_test` internals with a small
`tracing_subscriber` layer that records event fields into a buffer owned by the
scenario `World`.

Keep the tracing support in a test helper module so the BDD scenario file stays
below the 400-line limit.
