+++ tests/sse_wire_contract_bdd.rs
@@ -1,10 +1,6 @@
 //! Behavioural tests for the shared SSE wire contract.
 
-use std::{
-    fmt,
-    sync::{Arc, Mutex},
-    time::Duration,
-};
+use std::time::Duration;
 
 use actix_v2a::{
     CACHE_CONTROL_HEADER,
@@ -27,53 +23,11 @@ use actix_web::http::header::{HeaderMap, HeaderName, HeaderValue};
 use rstest::fixture;
 use rstest_bdd::Slot;
 use rstest_bdd_macros::{ScenarioState, given, scenario, then, when};
-use tracing_subscriber::{Layer, prelude::*};
 
-struct BufferLayer(Arc<Mutex<Vec<String>>>);
+#[path = "support/sse_trace_buffer.rs"]
+mod sse_trace_buffer;
 
-impl<S: tracing::Subscriber> Layer<S> for BufferLayer {
-    fn on_event(
-        &self,
-        event: &tracing::Event<'_>,
-        _ctx: tracing_subscriber::layer::Context<'_, S>,
-    ) {
-        use tracing_subscriber::field::Visit;
-
-        struct Collector(String);
-
-        impl Visit for Collector {
-            fn record_debug(&mut self, field: &tracing::field::Field, value: &dyn fmt::Debug) {
-                use fmt::Write as _;
-
-                let result = write!(self.0, " {}={:?}", field.name(), value);
-                debug_assert!(result.is_ok(), "writing to String should not fail");
-            }
-        }
-
-        let mut collector = Collector(String::new());
-        event.record(&mut collector);
-
-        if let Ok(mut buffer) = self.0.lock() {
-            buffer.push(collector.0.trim_start().to_owned());
-        }
-    }
-}
-
-struct TracingGuard(Mutex<Option<tracing::subscriber::DefaultGuard>>);
-
-impl fmt::Debug for TracingGuard {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter.write_str("TracingGuard(..)")
-    }
-}
-
-impl Drop for TracingGuard {
-    fn drop(&mut self) {
-        if let Ok(mut guard) = self.0.lock() {
-            let _ = guard.take();
-        }
-    }
-}
+use sse_trace_buffer::{LogBuffer, TracingGuard};
 
 #[derive(Debug, Default, ScenarioState)]
 struct World {
@@ -84,7 +38,7 @@ struct World {
     response_headers: Slot<Vec<SseHeader>>,
     heartbeat_frame: Slot<String>,
     event_frame: Slot<String>,
-    log_buffer: Slot<Arc<Mutex<Vec<String>>>>,
+    log_buffer: Slot<LogBuffer>,
     should_use_actix_path: Slot<bool>,
     stream_reset_frame: Slot<String>,
     tracing_guard: Slot<TracingGuard>,
@@ -92,14 +46,7 @@ struct World {
 
 impl World {
     fn install_tracing(&self) {
-        let buffer = Arc::new(Mutex::new(Vec::new()));
-        let subscriber = tracing_subscriber::registry().with(BufferLayer(Arc::clone(&buffer)));
-        let guard = TracingGuard(Mutex::new(Some(tracing::subscriber::set_default(
-            subscriber,
-        ))));
-
-        self.log_buffer.set(buffer);
-        self.tracing_guard.set(guard);
+        sse_trace_buffer::install_tracing(&self.log_buffer, &self.tracing_guard);
     }
 }
 
@@ -347,13 +294,7 @@ fn a_tracing_error_is_emitted_with_header_name_and_error_variant(
 }
 
 fn traced_scenario_logs_contain(world: &World, value: &str) -> bool {
-    let Some(buffer) = world.log_buffer.get() else {
-        return false;
-    };
-
-    buffer
-        .lock()
-        .is_ok_and(|lines| lines.iter().any(|line| line.contains(value)))
+    sse_trace_buffer::logs_contain(&world.log_buffer, value)
 }
 
 #[then("the response uses the canonical no-store cache policy")]
